// src/views/main/dashboard/analysis/type.ts
export interface InfoCardData {
  title: string
  number: string
  isAdd: boolean
  percentNumber: string
  icon: string
  time: string
  bottomColor: string
}
export interface ChartData {
  icon: string
  areaName: string
  incomeNumber: number
  percent: string
}

export interface TransactionData {
  img: string
  name: string
  date: string
  status: number
  type: string
  amount: number
}